Replace doubles by floats

+++ src/main/rotation/rotation.c
@@ -4,22 +4,22 @@
 #include <math.h>
 #include <stdlib.h>
 
-Matrix *rotate_matrix(const Matrix *src, double angle)
+Matrix *rotate_matrix(const Matrix *src, float angle)
 {
     if (src == NULL)
     {
         errx(EXIT_FAILURE, "rotate_matrix: src matrix is NULL");
     }
 
-    size_t w = (size_t)mat_width(src);
-    size_t h = (size_t)mat_height(src);
+    size_t w = mat_width(src);
+    size_t h = mat_height(src);
 
-    double rad = angle * M_PI / 180.0;
+    float rad = angle * M_PI / 180.0f;
 
     size_t nw =
-        (size_t)(fabs((double)w * cos(rad)) + fabs((double)h * sin(rad)) + 0.5);
+        (size_t)(fabsf((float)w * cosf(rad)) + fabsf((float)h * sinf(rad)) + 0.5f);
     size_t nh =
-        (size_t)(fabs((double)h * cos(rad)) + fabs((double)w * sin(rad)) + 0.5);
+        (size_t)(fabsf((float)h * cosf(rad)) + fabsf((float)w * sinf(rad)) + 0.5f);
 
     Matrix *rotated = mat_create_zero(nh, nw);
 
@@ -27,27 +27,27 @@ Matrix *rotate_matrix(const Matrix *src, double angle)
     {
         for (size_t x = 0; x < nw; x++)
         {
-            *mat_coef_ptr(rotated, y, x) = 255.0;
+            *mat_coef_ptr(rotated, y, x) = 255.0f;
         }
     }
 
-    double cx = w / 2.0;
-    double cy = h / 2.0;
-    double ncx = nw / 2.0;
-    double ncy = nh / 2.0;
+    float cx = (float)w / 2.0f;
+    float cy = (float)h / 2.0f;
+    float ncx = (float)nw / 2.0f;
+    float ncy = (float)nh / 2.0f;
 
     for (size_t y = 0; y < nh; y++)
     {
         for (size_t x = 0; x < nw; x++)
         {
-            double tx = (x - ncx) * cos(rad) - (y - ncy) * sin(rad) + cx;
-            double ty = (x - ncx) * sin(rad) + (y - ncy) * cos(rad) + cy;
+            float tx = ((float)x - ncx) * cosf(rad) - ((float)y - ncy) * sinf(rad) + cx;
+            float ty = ((float)x - ncx) * sinf(rad) + ((float)y - ncy) * cosf(rad) + cy;
 
-            if (tx >= 0 && tx < w && ty >= 0 && ty < h)
+            if (tx >= 0.0f && tx < (float)w && ty >= 0.0f && ty < (float)h)
             {
                 int ix = (int)tx;
                 int iy = (int)ty;
-                double val = mat_coef(src, iy, ix);
+                float val = mat_coef(src, iy, ix);
                 *mat_coef_ptr(rotated, y, x) = val;
             }
         }
@@ -56,16 +56,16 @@ Matrix *rotate_matrix(const Matrix *src, double angle)
     return rotated;
 }
 
-ImageData *rotate_image(ImageData *img, double angle)
+ImageData *rotate_image(ImageData *img, float angle)
 {
     int w = img->width;
     int h = img->height;
     Pixel *p = img->pixels;
 
-    double rad = angle * M_PI / 180.0;
+    float rad = angle * M_PI / 180.0;
 
-    int nw = (int)(fabs(w * cos(rad)) + fabs(h * sin(rad)) + 0.5);
-    int nh = (int)(fabs(h * cos(rad)) + fabs(w * sin(rad)) + 0.5);
+    int nw = (int)(fabsf((float)w * cosf(rad)) + fabsf((float)h * sinf(rad)) + 0.5f);
+    int nh = (int)(fabsf((float)h * cosf(rad)) + fabsf((float)w * sinf(rad)) + 0.5f);
 
     Pixel *np = calloc(nw * nh, sizeof(Pixel));
     if (!np)
@@ -80,17 +80,17 @@ ImageData *rotate_image(ImageData *img, double angle)
         np[i].b = 255;
     }
 
-    double cx = w / 2.0;
-    double cy = h / 2.0;
-    double ncx = nw / 2.0;
-    double ncy = nh / 2.0;
+    float cx = w / 2.0f;
+    float cy = h / 2.0f;
+    float ncx = nw / 2.0f;
+    float ncy = nh / 2.0f;
 
     for (int y = 0; y < nh; y++)
     {
         for (int x = 0; x < nw; x++)
         {
-            double tx = (x - ncx) * cos(rad) - (y - ncy) * sin(rad) + cx;
-            double ty = (x - ncx) * sin(rad) + (y - ncy) * cos(rad) + cy;
+            float tx = ((float)x - ncx) * cosf(rad) - ((float)y - ncy) * sinf(rad) + cx;
+            float ty = ((float)x - ncx) * sinf(rad) + ((float)y - ncy) * cosf(rad) + cy;
 
             if (tx >= 0 && tx < w && ty >= 0 && ty < h)
             {
